backend/requirements.txt
# Core Web Framework
fastapi>=0.95.1
uvicorn>=0.22.0
starlette>=0.26.1

# Authentication and Security
python-dotenv>=1.0.0
python-jose>=3.3.0
bcrypt>=4.0.1
authlib>=1.2.0
email-validator>=2.0.0
python-multipart>=0.0.9

# Machine Learning and AI
transformers>=4.35.2
torch>=2.1.0,<2.4.0
sentencepiece>=0.1.97
huggingface_hub>=0.19.4
numpy<2.0.0
safetensors>=0.4.1

# Data Validation
pydantic>=2.0.0

# HTTP and Networking
httpx>=0.24.1
requests>=2.31.0

# Text Processing
textblob>=0.17.1
openai>=1.3.7
langchain>=0.0.350
newspaper3k>=0.2.8
nltk>=3.8.1

# Optional Twitter Integration
tweepy>=4.14.0

# Development and Debugging
pytest>=7.3.1
typing-extensions>=4.5.0

# Database (if needed)
sqlalchemy>=1.4.46

# Additional Utilities
protobuf>=4.23.4
